requirements.txt
# Python dependencies
praw>=7.6.0
kafka-python>=2.0.2
tweepy>=4.12.0
numpy>=1.20.0
pandas>=1.3.0
nltk>=3.6.0
textblob>=0.15.3
scikit-learn>=1.0.0
torch>=1.10.0
transformers>=4.18.0
pyspark>=3.2.0
pyarrow>=6.0.0
matplotlib>=3.5.0
seaborn>=0.11.2
scipy>=1.7.0
statsmodels>=0.13.0
yfinance>=0.1.70
alpha_vantage>=2.3.1
streamlit>=1.20.0
plotly>=5.10.0
altair>=4.2.0

# New dependencies for advanced features
networkx>=2.8.0
langdetect>=1.0.9
python-louvain>=0.16.0
dash>=2.9.0
dash-bootstrap-components>=1.0.0
requests>=2.28.0
tqdm>=4.65.0

# Advanced analytics dependencies
gensim>=4.3.0
spacy>=3.5.0
pyldavis>=3.4.0
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
fasttext>=0.9.2
bertopic>=0.15.0
wordcloud>=1.8.2
pmdarima>=2.0.3
prophet>=1.1.4
vaex>=4.16.0
hdbscan>=0.8.29
umap-learn>=0.5.3
vaderSentiment>=3.3.2
contractions>=0.1.73
emoji>=2.2.0

# Core requirements
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0

# Data processing and text analysis
nltk>=3.6.0
spacy>=3.1.0
transformers>=4.10.0
torch>=1.9.0
wordcloud>=1.8.1
networkx>=2.6.0
newspaper3k>=0.2.8
feedparser>=6.0.0
requests>=2.26.0
beautifulsoup4>=4.10.0

# Streaming and real-time processing
kafka-python>=2.0.0
confluent-kafka>=1.7.0

# Database
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
yfinance>=0.1.63
alpha_vantage>=2.3.1

# Dashboard and visualization
streamlit>=1.8.0
plotly>=5.3.0
pydeck>=0.7.0

# API and web
fastapi>=0.70.0
uvicorn>=0.15.0
aiohttp>=3.8.0

# Misc
tqdm>=4.62.0
pytest>=6.2.5
black>=21.9b0
pyarrow>=6.0.0

# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.10.0

# Data analysis
scikit-learn>=1.0.0
nltk>=3.6.0
wordcloud>=1.8.1
textblob>=0.15.3
gensim>=4.1.2
statsmodels>=0.13.0

# Data collection
feedparser>=6.0.0
beautifulsoup4>=4.10.0
requests>=2.25.0
praw>=7.5.0
lxml>=4.6.0

# Database
sqlalchemy>=1.4.0

# Visualization
networkx>=2.6.0

# Optional: Advanced models
transformers>=4.12.0
torch>=1.10.0
